Move test helpers to helpers_test
So that they are easier to find without an IDE.
